Fix casting

+++ classes/class.ilNolejMediaSelectorGUI.php
@@ -462,7 +462,7 @@ class ilNolejMediaSelectorGUI
 
     public static function getSignedUrl($a_mob_id, bool $a_is_obj_id = false, int $ttl = 10): string
     {
-        $objId = $a_is_obj_id ? $a_mob_id : self::getObjId($a_mob_id);
+        $objId = (int) ($a_is_obj_id ? $a_mob_id : self::getObjId($a_mob_id));
         $path = ilObjMediaObject::_lookupItemPath($objId);
 
         $tokenMaxLifetimeInSeconds = ilWACSignedPath::getTokenMaxLifetimeInSeconds();
